requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-community==0.0.1
pyttsx3
SpeechRecognition
sympy
gtts
pydub
ffmpeg
pvporcupine
sounddevice
sentence-transformers
python-dotenv
PyAudio-0.2.14-cp39-cp39-win_amd64.whl